:pencil: | Float -> Double, pegar o ID da guild corretamente, guardar a instância do servidor e do serverConfig nas variáveis

+++ src/main/java/com/mrpowergamerbr/loritta/userdata/EconomyConfig.kt
@@ -12,5 +12,5 @@ class EconomyConfig {
 	var economyNamePlural: String? = null
 	// Taxa de transferência de Sonhos -> Local
 	@AllowReflection
-	var exchangeRate: Float? = null
+	var exchangeRate: Double? = null
 }
